rt_frtchord: handling of sticky and normal nodes
- sometimes, nodes are found between neighbours which haven't been added yet as sticky
nodes (e.g. because the periodic stabilization is pending). If this happens, such nodes
are added as sticky nodes instead, as they should become neighbours later.
- Former neighbors which are still alive should not be removed when updating the routing
table, but should be converted to normal routing table entries
- removed a wrong assertion in entry_filtering: when filtering nodes, the adjacent fingers
of sticky nodes can change, so asserting that the sticky nodes remained the same is fails