slide_chord: adjust rm subscription to indicate pred status correctly
This fixes occasionally hanging slides for jumping nodes. In the course
of a join, the node information saved in the slide_op record already has
the target id set as the node id. To trigger the ring maintenance
subscription when the pred changes, we need to use the node information
from the ring maintenance and not from the slide_op record.

This bug occured when the pred in the ring maintenance did not
immediately match the expected target id. Additionally, the pred's id
had to change to the correct id BEFORE the subscription in the ring
maintenance was established. Then, the ring maintenance subscription
would not trigger because it would wait for the pred to be different
from the node information provided in the slide_op record. Since this
node information would already be the established, slides would simply
starve.